+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Added constant ChangefeedFormatDynamoDBStreamsJSON for cdc format
+
 ## v3.48.8
 * Fixed `sugar.RemoveRecursive()` for column table type