+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Added `table/options.ChangefeedFormatDynamoDBStreamsJSON` format of `DynamoDB` change feeds
+
 ## v3.48.8
 * Fixed `sugar.RemoveRecursive()` for column table type