+++ CHANGELOG.md
@@ -1,4 +1,5 @@
-* Added `table.Session.ReadRows` 
+* Added `table.Session.ReadRows` method for getting rows by keys
+* Added `table/options.ChangefeedFormatDynamoDBStreamsJSON` format of `DynamoDB` change feeds
 
 ## v3.48.8
 * Fixed `sugar.RemoveRecursive()` for column table type